Extract Descriptor creation to addDescriptor method of the associated Relation.js type
